Develop

+++ theMovie-app/Models/Movie.swift
@@ -10,10 +10,12 @@ import Foundation
 
 struct Movie: Codable {
     
-    var id: Int
+    var id: Int64
     var title: String = ""
     var genreIds: [Int] = []
     var overview: String = ""
     var releaseDate: String?
     var posterPath: String
+    
+  
 }
